Untested LogDeleter changes
